% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bsr}
\alias{bsr}
\title{Run Hierarchichal Reconciliation with Bayesian Shrinkage}
\usage{
bsr(
  object,
  h,
  fmethod = "arima",
  burn_in = 10,
  length_sample = 100,
  shrinkage = NULL
)
}
\arguments{
\item{object}{gts or hts object.}

\item{h}{integer, forecasting horizon.}

\item{fmethod}{character, forecasting method to use ("arima", "ets" or "rw").
to shrink towards its base forecast.}

\item{burn_in}{integer, number of initial iterations to be discarded.}

\item{length_sample}{integer, sample size from joint posterior.}

\item{shrinkage}{character or numeric, indicates type shrinkage prior. Defaults to null.}
}
\value{
A list containing parameters for each horizon and gts output
}
\description{
BSR description
}
