% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{CollectOutput}
\alias{CollectOutput}
\title{Collect Output}
\usage{
CollectOutput(object, forecasts.list, results.list, pars)
}
\arguments{
\item{object}{gts object}

\item{results.list}{list of results}

\item{pars}{list of parameters}

\item{forecast.list}{list of forecasts}
}
\value{
gts objects
}
\description{
This function generates a gts object
}
