% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{create_predictions}
\alias{create_predictions}
\title{Define Weighting Matrix}
\usage{
create_predictions(object, fmethod, pars)
}
\arguments{
\item{object}{gts object}

\item{fmethod}{forecasting method to be used}

\item{pars}{list of parameters}
}
\value{
list of predictions
}
\description{
This function generates a diagonal matrix of weights used
}
