% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{CreatePredictions}
\alias{CreatePredictions}
\title{Create Predictions}
\usage{
CreatePredictions(object, fmethod, pars)
}
\arguments{
\item{object}{gts object}

\item{fmethod}{character, forecasting method to use ("arima", "ets" or "rw")}

\item{pars}{list of parameters}
}
\value{
diagonal sparse matrix
}
\description{
This function generates samples by drawing from fitted models
}
