% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{define_weights}
\alias{define_weights}
\title{Define Weighting Matrix}
\usage{
define_weights(pars)
}
\arguments{
\item{pars}{list of parameters}
}
\value{
vector of weights
}
\description{
This function generates a diagonal matrix of weights used
}
