% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{DefineWeights}
\alias{DefineWeights}
\title{Define Weighting Matrix}
\usage{
DefineWeights(S, pars)
}
\arguments{
\item{S}{S matrix}

\item{pars}{list of parameters}
}
\value{
diagonal sparse matrix
}
\description{
This function generates a diagonal matrix of weights used
}
