% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R, R/hello.R
\name{gg_run_cal}
\alias{gg_run_cal}
\title{Plot running calendar from Strava}
\usage{
gg_run_cal(data, year, current_year)

gg_run_cal(data, year, current_year)
}
\arguments{
\item{data}{a dataframe from Strava API}

\item{year}{current year to plot}

\item{current_year}{logical. Is year the current year.}
}
\value{
a ggplot

a ggplot
}
\description{
Plot running calendar from Strava

Plot running calendar from Strava
}
