% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{run_reconciliation}
\alias{run_reconciliation}
\title{Running Reconciliation}
\usage{
run_reconciliation(forecasts_list, pars)
}
\arguments{
\item{forecasts_list}{list of predictions}

\item{pars}{list of parameters}
}
\value{
matrix containing betas
}
\description{
This function estimates a Bayesian state space model to reconcile all
forecast horizons jointly
}
