% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfun.R
\name{RunReconciliation}
\alias{RunReconciliation}
\title{Reconcile Forecasts at each Horizon}
\usage{
RunReconciliation(S, forecasts.list, pars)
}
\arguments{
\item{object}{gts or hts object.}

\item{h}{integer, forecasting horizon.}

\item{fmethod}{character, forecasting method to use ("arima", "ets" or "rw").}

\item{series_to_be_shrunk}{vector of integers, indicating which reconciled
to shrink towards its base forecast.}

\item{nser_shr}{numerical, shrinks top level towards base forecasts.}

\item{xser_shr}{numerical, shrinks selected forecasts towards base forecasts.}
}
\value{
A list containing parameters for each horizon and gts output
}
\description{
This function approximates the joint posterior distribution using Gibbs sampling
}
