% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tradegts}
\alias{tradegts}
\title{Data on Exports of Swiss Goods}
\format{
Grouped Time Series
}
\source{
\href{https://www.ezv.admin.ch/ezv/de/home/themen/schweizerische-aussenhandelsstatistik.html}{Swiss Customs}
}
\usage{
data(tradegts)
}
\description{
Monthly time series from 1988-2018, not adjusted for seasonality and calendar effects,
containing regions and countries, main groups and subgroups
}
\keyword{datasets}
