% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tradegts_red}
\alias{tradegts_red}
\title{Reduced Data on Exports of Swiss Goods}
\format{
Grouped Time Series
}
\source{
\href{https://www.ezv.admin.ch/ezv/de/home/themen/schweizerische-aussenhandelsstatistik.html}{Swiss Customs}
}
\usage{
data(tradegts_red)
}
\description{
Monthly time series from 1988-2018, not adjusted for seasonality and calendar effects,
containing only regions and main groups
}
\keyword{datasets}
